requirements.txt
streamlit==1.40.0
openai==1.52.2
httpx==0.27.2
pyyaml==6.0.2
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3